separated code of login form and created hook useLogin with logic of the login form
