+++ client/src/components/forms/LoginForm/LoginForm.tsx
@@ -1,38 +1,17 @@
 import { Button, FormControl, FormHelperText, TextField } from "@mui/material"
-import { ChangeEvent, FormEvent, useState } from "react"
-import { login, selectError, selectLoading } from "../../../features/authentication/authenticationSlice"
-import { useAppDispatch, useAppSelector } from "../../../app/hooks"
-
-const generalProps = {
-    size : "small" as "small",
-    required : true,
-    margin : "normal" as "normal"
-}
+import { ChangeEvent } from "react"
+import { inputConfig } from "../globalProps"
+import useLogin from "../../../hooks/useLogin"
 
 export default function LoginForm(){
-    const [fields, setFields] = useState({
-        username : "",
-        password : "",
-    })
-    const error = useAppSelector(selectError)
-    const isLoading = useAppSelector(selectLoading)
-    const dispatch = useAppDispatch()
-
-    const onChange = (event: ChangeEvent<HTMLInputElement>, field: "username" | "password")=>{
-        setFields(fields=>({...fields, [field] : event.target.value}))
-    }
-
-    const onSubmit = (event: FormEvent<HTMLFormElement>)=>{
-        event.preventDefault()
-        dispatch(login(fields))
-    }
+    const { fields, onChange, onSubmit, error, isLoading } = useLogin()
 
     return (
         <form onSubmit={onSubmit}>
             <FormControl sx={{ padding : "50px" }}>
                 <TextField
                     label="Username"
-                    {...generalProps}
+                    {...inputConfig}
                     value={fields.username}
                     error={Boolean(error)}
                     onChange={(e:ChangeEvent<HTMLInputElement>)=>onChange(e, "username")}
@@ -41,7 +20,7 @@ export default function LoginForm(){
                     label="Password"
                     type="password"
                     error={Boolean(error)}
-                    {...generalProps}
+                    {...inputConfig}
                     value={fields.password}
                     onChange={(e:ChangeEvent<HTMLInputElement>)=>onChange(e, "password")}
                 />
