+++ client/src/components/forms/globalProps.ts
@@ -1,0 +1,5 @@
+export const inputConfig = {
+    size : "small" as "small",
+    required : true,
+    margin : "normal" as "normal"
+}
